+++ src/main/java/persistence/sql/QueryBuilder.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 public interface QueryBuilder {
