+++ src/main/java/persistence/sql/dialect/h2/H2ColumnTypeProperties.java
@@ -1,10 +1,8 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
 import jakarta.persistence.Column;
 
 import java.lang.reflect.Field;
-
-import static persistence.sql.ddl.dialect.h2.H2Query.*;
 
 /**
  * 컬럼 프로퍼티 쿼리 생성을 위한 클래스
@@ -15,9 +13,9 @@ public class H2ColumnTypeProperties {
     public static String getVarcharLength(Field entityField) {
         Column columnAnnotation = entityField.getDeclaredAnnotation(Column.class);
         if (columnAnnotation == null) {
-            return OPEN_PARENTHESIS + VARCHAR_DEFAULT_LENGTH + CLOSE_PARENTHESIS;
+            return H2Query.OPEN_PARENTHESIS + H2Query.VARCHAR_DEFAULT_LENGTH + H2Query.CLOSE_PARENTHESIS;
         }
-        return OPEN_PARENTHESIS + entityField.getDeclaredAnnotation(Column.class).length() + CLOSE_PARENTHESIS;
+        return H2Query.OPEN_PARENTHESIS + entityField.getDeclaredAnnotation(Column.class).length() + H2Query.CLOSE_PARENTHESIS;
     }
 
 }
