+++ src/main/java/persistence/sql/dialect/h2/H2CreateQuery.java
@@ -1,9 +1,9 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
-import persistence.sql.ddl.entity.EntityColumn;
-import persistence.sql.ddl.entity.EntityData;
+import persistence.sql.entity.EntityColumn;
+import persistence.sql.entity.EntityData;
 
 import java.util.List;
