+++ src/main/java/persistence/sql/dialect/h2/H2Dialect.java
@@ -1,7 +1,7 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
-import persistence.sql.ddl.dialect.Dialect;
-import persistence.sql.ddl.entity.EntityData;
+import persistence.sql.Dialect;
+import persistence.sql.entity.EntityData;
 
 /**
  * H2 DB 쿼리 생성
