+++ src/test/java/persistence/sql/dialect/h2/H2ColumnTypePropertiesTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
