+++ src/test/java/persistence/sql/dialect/h2/H2CreateQueryTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.dialect.h2;
+package persistence.sql.dialect.h2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -7,7 +7,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import persistence.entity.Person_Week1_Step2_Demand1;
 import persistence.entity.Person_Week1_Step2_Demand2;
 import persistence.entity.Person_Week1_Step2_Demand3;
-import persistence.sql.ddl.entity.EntityData;
+import persistence.sql.entity.EntityData;
 
 import java.util.stream.Stream;
